Arreglar Rebuig i invalidació unitaries. Taula de tests de tamany de PDF. API Indra taula de firmes al final.
git-svn-id: https://svn.code.sf.net/p/portafib/code-0/portafib-core/branches/portafib-1.0@117 1a83b7bb-df65-4973-8540-7514423c1e25

+++ ws/portafib_api_example/src/main/java/es/caib/portafib/ws/v1/example/Exemple.java
@@ -103,7 +103,8 @@ public class Exemple {
       log.info("Versio " + api.getVersion());
       log.info("Idiomes:" + Arrays.toString(api.getSupportedLanguages().toArray()));
 
-      final String titol = "Peticio de Test amb Custodia";
+      final String titol = "Peticio de Test amb Custodia - "
+         + ((System.currentTimeMillis() / 1000 ) % 100000);
       final String remitent = "Helium";
 
       // Existeix sistema de custodia
@@ -211,7 +212,7 @@ public class Exemple {
         }
 
       } finally {
-        if (peticioDeFirmaID != null) {
+        if (peticioDeFirmaID != null && isDeleteOnFinish()) {
           // Borrar la petició
           api.deletePeticioDeFirma(peticioDeFirmaID);
         }
@@ -259,6 +260,10 @@ public class Exemple {
   
   public static boolean isWaitToSign() {
     return "true".equals(testProperties.getProperty("waittosign"));
+  }
+  
+  public static boolean isDeleteOnFinish() {
+    return "true".equals(testProperties.getProperty("deleteonfinish"));
   }
